FIX: 32-bit support in uinput for ARMv7 Linux

+++ src/linux_wayland/ffi/input.rs
@@ -4,8 +4,8 @@
 #[allow(non_camel_case_types)]
 pub struct timeval {
     // Not quite sure if these should be 32 or 64
-    pub tv_sec: i64,
-    pub tv_usec: i64,
+    pub tv_sec: i32,
+    pub tv_usec: i32,
 }
 
 #[repr(C)]
